rtl: Fix CRC E bit generation
According to the spec, the CRC E bits are 1=OK, 0=Error and they
must also be transmitted as 0 if RX synchronization is not achieved.

To stay consistent everywhere, the status reported in the RX BD
status, the bits are also defined as 1=OK, 0=Error and the doc is
updated to reflect that.

This fixes issue OS#4661

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>